libs/requirements.txt
# Shared libraries requirements
# These dependencies are used by the common libraries

# Authentication and security
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
cryptography==45.0.6

# HTTP and API
fastapi>=0.104.1
starlette>=0.27.0
httpx==0.25.2

# Database
sqlalchemy>=2.0.0
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Cache and messaging
redis>=5.0.1
paho-mqtt==1.6.1

# Monitoring and observability
psutil==5.9.6

# Utilities
python-dateutil==2.8.2
pydantic>=2.6.0,<2.10.0
pydantic-settings>=2.1.0
